requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
google-generativeai
deep-translator
edge-tts
mutagen
